Added Server-Side querying for database objects
