+++ next.config.js
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  experimental:{
+    serverActions: true,
+  }
+};
 
 module.exports = nextConfig;
