+++ src/model/Column/ColumnDatabase.ts
@@ -1,0 +1,8 @@
+interface ColumnEntity {
+  id: string;
+  name: string;
+  position: number;
+  color: string;
+}
+
+export const columnDatabase = new Map<string, ColumnEntity>();
