+++ src/model/Column/ColumnRepository.ts
@@ -1,15 +1,7 @@
-import { AbstractRepository } from "@/model/AbstractRepository";
+import { AbstractRepository, Column } from "@/model/types";
+import { columnDatabase } from "./ColumnDatabase";
 
-export interface ColumnEntity {
-  id: string;
-  name: string;
-  position: number;
-  color: string;
-}
-
-const columnDatabase = new Map<string, ColumnEntity>();
-
-export class ColumnRepository implements AbstractRepository<ColumnEntity> {
+export class ColumnRepository implements AbstractRepository<Column> {
   async findById(id: string) {
     const item = columnDatabase.get(id);
     if (!item) {
@@ -23,12 +15,12 @@ export class ColumnRepository implements AbstractRepository<ColumnEntity> {
     return Array.from(columnDatabase.values());
   }
 
-  async create(column: ColumnEntity) {
+  async create(column: Column) {
     columnDatabase.set(column.id, column);
     return column;
   }
 
-  async update(id: string, column: ColumnEntity) {
+  async update(id: string, column: Column) {
     columnDatabase.set(id, column);
     return column;
   }
