+++ src/model/Column/index.ts
@@ -1,0 +1,1 @@
+export { ColumnRepository } from "./ColumnRepository";
