+++ src/model/Task/TaskDatabase.ts
@@ -1,0 +1,8 @@
+interface TaskEntity {
+  id: string;
+  text: string;
+  columnId: string;
+  position: number;
+}
+
+export const taskDatabase = new Map<string, TaskEntity>();
