+++ src/model/Task/index.ts
@@ -1,0 +1,1 @@
+export { TaskRepository } from "./TaskRepository";
